# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: RouteLoom API
  version: 0.1.0
  description: Private Responses-compatible and native Jobs API for quota-aware model routing
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: https://router.tailnet.example
    description: Private Tailnet endpoint
security:
  - bearerAuth: []
paths:
...
  /auth/passkey/register/options:
    post:
      operationId: beginPasskeyRegistration
      security: []
      parameters:
        - $ref: "#/components/parameters/BootstrapToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                displayName:
                  type: string
      responses:
        "201":
          description: WebAuthn registration challenge and browser options
  /auth/passkey/register/verify:
    post:
      operationId: verifyPasskeyRegistration
      security: []
      parameters:
        - $ref: "#/components/parameters/BootstrapToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [challengeId, response]
              properties:
                challengeId:
                  type: string
                  format: uuid
                response:
                  type: object
      responses:
        "201":
          description: Registration result and ten one-time recovery codes
  /auth/passkey/login/options:
    post:
      operationId: beginPasskeyLogin
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        "201":
          description: WebAuthn authentication challenge and browser options
  /auth/passkey/login/verify:
    post:
      operationId: verifyPasskeyLogin
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [challengeId, response]
              properties:
                challengeId:
                  type: string
                  format: uuid
                response:
                  type: object
      responses:
        "201":
          description: Authentication result with an HttpOnly session cookie
  /auth/recovery/login:
    post:
      operationId: loginWithRecoveryCode
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, code]
              properties:
                email:
                  type: string
                  format: email
                code:
                  type: string
      responses:
        "201":
          description: Authentication result after consuming one recovery code
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: AIALRA_API_KEY
  parameters:
...
    BootstrapToken:
      name: X-Bootstrap-Token
      in: header
      required: true
      schema:
        type: string
        minLength: 32
...